+++ packages/super-editor/src/extensions/passthrough/passthrough.js
@@ -36,6 +36,7 @@ export const PassthroughInline = Node.create({
   name: 'passthroughInline',
   group: 'inline',
   inline: true,
+  marks: '',
   // IMPORTANT: This node is registered in pm-adapter/src/constants.ts ATOMIC_INLINE_TYPES
   // If you change atom to false, you MUST remove it from that set to avoid positioning bugs
   atom: true,